+++ src/main/java/uk/gov/hmcts/reform/orgrolemapping/data/AccessTypesEntity.java
@@ -8,28 +8,19 @@ import lombok.Setter;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.SequenceGenerator;
-import java.io.Serial;
-import java.io.Serializable;
+
 
 @Builder(toBuilder = true)
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@SequenceGenerator(name = "version_seq", sequenceName = "version_seq", allocationSize = 1)
 @Entity(name = "access_types")
-public class AccessTypesEntity implements Serializable {
-
-    @Serial
-    private static final long serialVersionUID = 1L;
+public class AccessTypesEntity  {
 
     @Id
     @Column(name = "version")
-    @GeneratedValue(strategy = GenerationType.IDENTITY, generator = "version_seq")
     private Long version;
 
     @Column(name = "access_types", nullable = false)
