+++ src/main/java/uk/gov/hmcts/reform/orgrolemapping/domain/model/OrganisationInfo.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import uk.gov.hmcts.reform.orgrolemapping.domain.OrganisationStatus;
+import uk.gov.hmcts.reform.orgrolemapping.domain.model.enums.OrganisationStatus;
 
 import java.time.LocalDateTime;
 import java.util.List;
