+++ src/main/java/uk/gov/hmcts/reform/orgrolemapping/domain/model/UserAccessType.java
@@ -1,0 +1,19 @@
+package uk.gov.hmcts.reform.orgrolemapping.domain.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Setter
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class UserAccessType {
+
+    private String jurisdictionId;
+    private String organisationProfileId;
+    private String accessTypeId;
+    private Boolean enabled;
+
+}
